Add schema decorator factory function and start changing relation
